refactor(web): mueve las paginas personales a /dashboard/*
