+++ apps/web/src/app/dashboard/notifications/mark-all-read-button.tsx
