+++ apps/web/src/components/molecules/notification-item.tsx
@@ -2,7 +2,7 @@
 
 import Link from 'next/link';
 import { useTransition } from 'react';
-import { markNotificationReadAction } from '@/app/panel/notificaciones/actions';
+import { markNotificationReadAction } from '@/app/dashboard/notifications/actions';
 import { useLocale } from '@/i18n/locale-context';
 import { getMessages } from '@/i18n';
 import { LocalDate } from '@/components/atoms/local-date';
